requirements.txt
amqp==5.0.6
asgiref==3.4.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
cheroot==8.5.2
CherryPy==18.6.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.4
django-admin-tools==0.9.1
django-cors-headers==3.7.0
django-currentuser==0.5.3
django-filter==2.4.0
django-modeltranslation==0.17.3
django-simple-history==3.0.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
drf-base64==2.0
et-xmlfile==1.1.0
fcm-django==0.3.10
glob2==0.7
idna==2.10
importlib-metadata==4.6.0
importlib-resources==5.2.0
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.text==3.5.0
jasmine-core==3.7.1
Jinja2==2.11.3
kombu==5.1.0
MarkupSafe==2.0.1
more-itertools==8.8.0
munch==2.5.0
openpyxl==3.0.7
ordereddict==1.1
Pillow==8.2.0
portend==2.7.1
prompt-toolkit==3.0.19
pyfcm==1.4.9
PyJWT==2.1.0
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
sendgrid==6.7.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
tempora==4.1.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zc.lockfile==2.0
zipp==3.4.1
